core: Rename patcher file
