+++ src/core/arm/nce/patcher.cpp
@@ -7,7 +7,7 @@
 #include "core/arm/nce/arm_nce.h"
 #include "core/arm/nce/guest_context.h"
 #include "core/arm/nce/instructions.h"
-#include "core/arm/nce/patch.h"
+#include "core/arm/nce/patcher.h"
 #include "core/core.h"
 #include "core/core_timing.h"
 #include "core/hle/kernel/svc.h"
