+++ src/core/arm/nce/patcher.h
