+++ src/core/loader/nro.cpp
@@ -23,7 +23,7 @@
 #include "core/memory.h"
 
 #ifdef HAS_NCE
-#include "core/arm/nce/patch.h"
+#include "core/arm/nce/patcher.h"
 #endif
 
 namespace Loader {
