+++ src/core/loader/nso.cpp
@@ -21,7 +21,7 @@
 #include "core/memory.h"
 
 #ifdef HAS_NCE
-#include "core/arm/nce/patch.h"
+#include "core/arm/nce/patcher.h"
 #endif
 
 namespace Loader {
